refactor: notificationManager에게 메세지를 던지도록 수정
